refactor: change KC controller package name
Old package name was messing with imports.
