--- src/Tutor.Web/Controllers/KnowledgeComponent/KCProfile.cs
@@ -1,13 +1,0 @@
-﻿using AutoMapper;
-using Tutor.Web.Controllers.KnowledgeComponent.DTOs;
-
-namespace Tutor.Web.Controllers.KnowledgeComponent
-{
-    public class KCProfile : Profile
-    {
-        public KCProfile()
-        {
-            CreateMap<Core.KnowledgeComponentModel.KnowledgeComponents.KnowledgeComponent, KnowledgeComponentDTO>();
-        }
-    }
-}
